+++ Cudheart/Testing/TestUtil.cu
@@ -15,6 +15,7 @@ namespace Cudheart::Testing {
 
 	Testing::Testing() {
 		m_code = "# this file is generated automatically to simplify unit testing :)\n\nfrom util import *\n\nimport warnings\n\nwarnings.filterwarnings('ignore', category = DeprecationWarning)\n\n";
+		m_code += "# anti future headache tech\ntrue = True\nfalse = False\n\n";
 	}
 
 	Testing* Testing::get() {