+++ Changelog.md
@@ -47,6 +47,9 @@
 - fixed argwhere
 - fixed searchsorted
 - fixed some sorting mishaps
+- added some overloads for allclose
+- added isclose
+- changed some logic function's output
 
 ## notes:
 -